python_processor/requirements.txt
kafka-python
requests
clickhouse-driver
